pyproject.toml
[project]
name = "hackathon-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.9.0",
    "aiofiles>=24.1.0",
    "aiosqlite>=0.20.0",
    "azure-ai-textanalytics>=5.3.0",
    "azure-cognitiveservices-speech>=1.45.0",
    "azure-data-tables>=12.7.0",
    "azure-identity>=1.23.1",
    "azure-storage-blob>=12.26.0",
    "datasets>=4.0.0",
    "deepface>=0.0.93",
    "fastapi>=0.116.1",
    "librosa>=0.11.0",
    "neurokit2>=0.2.12",
    "numpy>=2.1.3",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-socketio>=5.13.0",
    "scipy>=1.16.0",
    "tensorflow>=2.19.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "transformers>=4.54.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
